Fix large size segments issues.

+++ SegmentDownloader.Core/UI/Connection.cs
@@ -63,7 +63,7 @@ namespace SegmentDownloader.Core.UI
 
         private void UpdateControls()
         {
-            lblMinSize.Text = ByteFormatter.ToString((int)numMinSegSize.Value);
+            lblMinSize.Text = ByteFormatter.ToString((long)numMinSegSize.Value);
         }
     }
 }
